#2197 Prevent creation of renaming commands if not necessary
